+++ packages/rx_bloc_list/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [4.0.0] - September 17, 2021
+* [BREAKING] Make onBottomScrolled optional parameter.
+* [BREAKING] Add a required parameter `loadNextPage`.
+ - Callback triggered once the user gets to the bottom of the list and there are more records to load.
+
 ## [3.0.1] - July 21, 2021
 * `PaginatedList`.`reset(hard: true)` now triggers `RxPaginatedBuilder`.`buildLoading`